+++ SensateIoT.Platform.Network.StorageService/Config/InternalBrokerConfig.cs
@@ -5,8 +5,11 @@
  * @email  michel@michelmegens.net
  */
 
+using JetBrains.Annotations;
+
 namespace SensateIoT.Platform.Network.StorageService.Config
 {
+	[UsedImplicitly]
 	public class InternalBrokerConfig
 	{
 		public string Username { get; set; }
@@ -16,5 +19,7 @@ namespace SensateIoT.Platform.Network.StorageService.Config
 		public short Port { get; set; }
 		public string BulkMeasurementTopic { get; set; }
 		public string BulkMessageTopic { get; set; }
+		public string TriggerEventQueueTopic { get; set; }
+		public string NetworkEventQueueTopic { get; set; }
 	}
 }