+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Updated
 - Statistics logging
 - Logging in the Storage Service
+- Router initialization routine
+- The database function `generic_getblobs`
 
 ### Removed
 - Unused code